Refactor sliders
