+++ src/Destinations/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Slider from "../Slider";
+import { Slider } from "../UI/Slider";
 import Card from "./Card";
 
 import data from "./staticData";
