+++ src/UI/Slider/index.js
@@ -1,0 +1,19 @@
+import React from "react";
+import styled from "styled-components";
+
+import { SliderContainer, SliderRow, NextBtn } from "./styled";
+
+export const Slide = styled.div`
+  flex: 0 0 auto;
+  margin: 0 8px;
+`;
+
+export const Slider = function({ children, hasDesktopNavigation }) {
+  return (
+    <SliderContainer>
+      {/* TODO: PrevBtn */}
+      <SliderRow>{children}</SliderRow>
+      {hasDesktopNavigation && <NextBtn />}
+    </SliderContainer>
+  );
+};
